WIP: assume future generators are never poisoned
`make_future()` is unsafe, thus we can be assured that the generator
constructor will never panic when called. The macro also generates the
code that calls it anyway.

NOTE: This actually probably isn't safe yet due to the use of ergo_pin,
which will probably put a panic in the wrong place if asked nicely.